Add files via upload
I revised the bugs of Andrew's code and rewrote some functions in it. It may contain other bugs because I haven't tested it but you can still get my basic idea. Yesterday Andrew and I talked about dealing with the windows in which each genotype has at least 2 hetero by using that geno with 2 hetero as seeds. There could then be different haplo we get from this geno. We then compare the total number of haplo generated in this window to choose the haplos we want. I am going to implement it today too.